requirements.txt
customtkinter
yt-dlp
pillow
requests
pywebview
